RNG for the inverse gaussian distribution

Based on code from simplexreg Peng Zhang, Zhenguo Qiu, Chengchun Shi (2016). simplexreg: An R Package for Regression Analysis of Proportional Data Using the Simplex Distribution. Journal of Statistical Software, 71(11), 1-21. doi:10.18637/jss.v071.i11

Usage

rIG(epsilon, Tau)

Arguments

epsilon

epsilon parameter of the inverse gaussian distribution

Tau

tau parameter of the inverse gaussian distribution

Value

a single sample from the specified inverse gaussian